Moving editbatch form to MVVM
