+++ ActivityScheduler.csproj
@@ -45,10 +45,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="Core\SetParentActivities.xaml.cs">
+    <Compile Update="Gui\SetParentActivitiesWindow\SetParentActivities.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Update="Core\EditBatch.xaml.cs">
+    <Compile Update="Gui\EditWindow\EditBatch.xaml.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Update="Properties\Resources.Designer.cs">
@@ -63,6 +63,10 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Core\Settings\" />
   </ItemGroup>
 
 </Project>
