--- Core/EditBatch.xaml
@@ -1,59 +1,0 @@
-﻿<Window x:Name="EditBatchFrm" x:Class="ActivityScheduler.Core.EditBatch"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:ActivityScheduler.Core" xmlns:wpfx="http://schemas.xceed.com/wpf/xaml/toolkit"
-        mc:Ignorable="d"
-        Title="Edit Batch" Height="651" Width="1224"
-        ResizeMode="NoResize" Loaded="Window_Loaded">
-    <Grid HorizontalAlignment="Center" Width="1224">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-        <Button Content="Save batch" Name="BatchSave" HorizontalAlignment="Left" Margin="1105,13,0,0" VerticalAlignment="Top" Width="99" Click="BatchSave_Click"  />
-        <Button Content="Create Activity" Name="CreateActivity" HorizontalAlignment="Left" Margin="701,45,0,0" VerticalAlignment="Top" Width="99" Click="CreateActivity_Click"  />
-        <Button Content="Cancel" x:Name="Cancel" HorizontalAlignment="Left" Margin="1105,43,0,0" VerticalAlignment="Top" Width="99" Click="Cancel_Click" />
-
-        <TextBlock HorizontalAlignment="Left" Name="BatchNumberLabel" Margin="28,24,0,0" TextWrapping="Wrap" Text="Number" VerticalAlignment="Top" />
-        <TextBox HorizontalAlignment="Left" Name="BatchNumberTb" Margin="28,45,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="78"/>
-
-        <TextBlock HorizontalAlignment="Left" Name="BatchNameLabel" Margin="125,24,0,0" TextWrapping="Wrap" Text="Name" VerticalAlignment="Top" />
-        <TextBox HorizontalAlignment="Left" Name="BatchNameTb" Margin="125,45,0,0" TextWrapping="Wrap" Text="" VerticalAlignment="Top" Width="155" />
-
-        <DataGrid HorizontalAlignment="Left" Name="ActivityGrid" Height="519" Margin="28,88,0,0" VerticalAlignment="Top" Width="772" SelectionMode="Single" IsReadOnly="True" CanUserAddRows="False" SelectionChanged="ActivityGrid_SelectionChanged"/>
-        <Canvas Name ="ActivityEditCanvas" Margin="815,88,20,210" Background="{DynamicResource {x:Static SystemColors.InactiveCaptionBrushKey}}">
-
-            <TextBlock TextWrapping="Wrap" Text="Activity Id" HorizontalAlignment="Left" Canvas.Left="10" Canvas.Top="21" VerticalAlignment="Top" />
-            <wpfx:MaskedTextBox TextWrapping="Wrap" x:Name="ActivityIdTb" Mask="000" Text="" Width="57" HorizontalAlignment="Center" Canvas.Left="101" Canvas.Top="21" VerticalAlignment="Top"  TextChanged="ActivityIdTb_TextChanged" Height="22"/>
-
-            <TextBlock TextWrapping="Wrap" Text="Name" Canvas.Left="10" Canvas.Top="48" HorizontalAlignment="Center" VerticalAlignment="Top"/>
-            <TextBox TextWrapping="Wrap" Name="ActivityNameTb" Text="" Width="256" Canvas.Left="101" Canvas.Top="48" HorizontalAlignment="Left" VerticalAlignment="Center" Height="19"/>
-
-            <CheckBox Content="Is hub" HorizontalAlignment="Center" VerticalAlignment="Center" Width="89" Name="IsHub" Canvas.Left="101" Canvas.Top="77"/>
-
-            <TextBlock TextWrapping="Wrap" Text="TransactionId" Canvas.Left="13" Canvas.Top="106" HorizontalAlignment="Left" VerticalAlignment="Top"/>
-            <wpfx:MaskedTextBox Width="92" Height="18" Mask="00:00:00" Canvas.Left="102" Canvas.Top="157" HorizontalAlignment="Left" VerticalAlignment="Center" x:Name="StartTimeTb" LostFocus="StartTimeTb_LostFocus" />
-
-            <TextBlock TextWrapping="Wrap" Text="Start time" Canvas.Left="11" Canvas.Top="159" HorizontalAlignment="Center" VerticalAlignment="Top"/>
-            <CheckBox Content="Is domestic" HorizontalAlignment="Left" Margin="0,0,0,0" VerticalAlignment="Top" Width="89" Name="IsDomestic" Canvas.Left="101" Canvas.Top="134"/>
-
-            <Button Content="Save Activity" Name="SaveActivityBtn" HorizontalAlignment="Center" VerticalAlignment="Top" Width="99"  Grid.Column="2" Grid.ColumnSpan="2" Canvas.Left="13" Canvas.Top="301" Click="SaveActivityBtn_Click"  />
-            <wpfx:MaskedTextBox Width="93" Height="19" Mask="000000" Canvas.Left="101" Canvas.Top="102" HorizontalAlignment="Left" VerticalAlignment="Center" x:Name="TransactionIdTb"/>
-
-            <CheckBox Content="Always success" HorizontalAlignment="Left" VerticalAlignment="Center" Width="120" Name="AlwaysSuccess" Canvas.Left="101" Canvas.Top="203"/>
-
-            <TextBox TextWrapping="Wrap" Name="ParentActivitiesTb" Text="" Width="221" Canvas.Left="102" Canvas.Top="231" HorizontalAlignment="Left" VerticalAlignment="Center" Height="19" IsReadOnly="True"/>
-            <TextBlock TextWrapping="Wrap" Text="Parent activities" Canvas.Left="11" Canvas.Top="233" HorizontalAlignment="Center" VerticalAlignment="Top"/>
-            <Button Content=">>" Canvas.Left="328" Canvas.Top="231" Name="SetParentActivitiesBtn" Click="SetParentActivitiesBtn_Click"/>
-            <Button Content="Delete activity" Name="DeleteActivityBtn" HorizontalAlignment="Left" VerticalAlignment="Top" Width="99" Canvas.Left="280" Canvas.Top="301" Background="#FFF90202" Margin="0,0,0,0" Click="DeleteActivityBtn_Click"  />
-
-
-
-        </Canvas>
-        <TextBlock HorizontalAlignment="Left" Margin="815,15,0,0" VerticalAlignment="Top" Width="285" Name="MsgLabel" Foreground="#FFF90202" Height="48"/>
-
-
-    </Grid>
-
-</Window>
