+++ Gui/SetParentActivitiesWindow/SetParentActivities.xaml.cs
