+++ Gui/SettingsWindow/Settings.xaml.cs
