+++ Gui/TrayContextMenu.xaml.cs
